Splits summaries in to module & refactors
